Move logger from all services to the platform package
