+++ platform/go.mod
@@ -1,0 +1,8 @@
+module github.com/ivyoverflow/pub-sub/platform
+
+go 1.15
+
+require (
+	go.uber.org/multierr v1.6.0 // indirect
+	go.uber.org/zap v1.16.0
+)
