--- server/cover.out
@@ -1,18 +1,0 @@
-mode: set
-github.com/ivyoverflow/pub-sub/server/internal/service/pub_sub.go:13.52,18.2 3 1
-github.com/ivyoverflow/pub-sub/server/internal/service/pub_sub.go:21.75,25.41 3 1
-github.com/ivyoverflow/pub-sub/server/internal/service/pub_sub.go:25.41,26.37 1 1
-github.com/ivyoverflow/pub-sub/server/internal/service/pub_sub.go:26.37,28.4 1 1
-github.com/ivyoverflow/pub-sub/server/internal/service/pub_sub.go:33.73,41.2 5 1
-github.com/ivyoverflow/pub-sub/server/internal/handler/subscriber.go:22.86,24.2 1 1
-github.com/ivyoverflow/pub-sub/server/internal/handler/subscriber.go:27.58,28.6 1 1
-github.com/ivyoverflow/pub-sub/server/internal/handler/subscriber.go:28.6,30.62 2 1
-github.com/ivyoverflow/pub-sub/server/internal/handler/subscriber.go:37.3,39.37 3 0
-github.com/ivyoverflow/pub-sub/server/internal/handler/subscriber.go:30.62,35.4 3 1
-github.com/ivyoverflow/pub-sub/server/internal/handler/subscriber.go:39.37,40.33 1 0
-github.com/ivyoverflow/pub-sub/server/internal/handler/subscriber.go:40.33,45.62 2 0
-github.com/ivyoverflow/pub-sub/server/internal/handler/subscriber.go:45.62,50.6 3 0
-github.com/ivyoverflow/pub-sub/server/internal/handler/publisher.go:21.84,23.2 1 1
-github.com/ivyoverflow/pub-sub/server/internal/handler/publisher.go:26.76,29.65 3 1
-github.com/ivyoverflow/pub-sub/server/internal/handler/publisher.go:36.2,37.136 2 1
-github.com/ivyoverflow/pub-sub/server/internal/handler/publisher.go:29.65,34.3 3 1
