+++ server/go.mod
@@ -3,9 +3,9 @@ module github.com/ivyoverflow/pub-sub/server
 go 1.15
 
 require (
+	github.com/ivyoverflow/pub-sub/platform v0.0.0-00010101000000-000000000000
 	github.com/pkg/errors v0.9.1 // indirect
 	github.com/stretchr/testify v1.5.1
-	go.uber.org/zap v1.16.0
 	golang.org/x/lint v0.0.0-20200302205851-738671d3881b // indirect
 	golang.org/x/net v0.0.0-20201224014010-6772e930b67b
 	golang.org/x/tools v0.0.0-20200818005847-188abfa75333 // indirect
@@ -13,3 +13,5 @@ require (
 	gopkg.in/yaml.v2 v2.2.8 // indirect
 	honnef.co/go/tools v0.0.1-2020.1.4 // indirect
 )
+
+replace github.com/ivyoverflow/pub-sub/platform => ../platform
