+++ server/internal/handler/publisher.go
@@ -6,7 +6,7 @@ import (
 	"fmt"
 	"net/http"
 
-	"github.com/ivyoverflow/pub-sub/server/internal/logger"
+	"github.com/ivyoverflow/pub-sub/platform/logger"
 	"github.com/ivyoverflow/pub-sub/server/internal/model"
 	"github.com/ivyoverflow/pub-sub/server/internal/service"
 )
