+++ server/internal/handler/subscriber.go
@@ -7,7 +7,7 @@ import (
 
 	"golang.org/x/net/websocket"
 
-	"github.com/ivyoverflow/pub-sub/server/internal/logger"
+	"github.com/ivyoverflow/pub-sub/platform/logger"
 	"github.com/ivyoverflow/pub-sub/server/internal/model"
 	"github.com/ivyoverflow/pub-sub/server/internal/service"
 )
